targuetero "Diseñador de tarjetas con QR" implementado completamente, 5 modelos, carga fotografia, llena datos personales, limpia la tarjeta, imprime la tarjeta.
