+++ PublicTransport/nbproject/project.properties
@@ -59,7 +59,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=publictransport.PublicTransport
+main.class=Interface.PublicTransport
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
