+++ PublicTransport/src/Interface/PublicTransport.java
@@ -3,8 +3,9 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package publictransport;
+package Interface;
 
+import Interface.Interface;
 import java.util.Date;
 
 /**
